Add `User-Agent` header
The format of the user agent is:

```
Parsec-Client/{version}; {platform}
```
