+++ libparsec/crates/client_connection/src/anonymous_cmds.rs
@@ -39,7 +39,7 @@ impl AnonymousCmds {
         proxy: ProxyConfig,
     ) -> anyhow::Result<Self> {
         let client = {
-            let builder = reqwest::ClientBuilder::default();
+            let builder = reqwest::ClientBuilder::default().user_agent(crate::CLIENT_USER_AGENT);
             let builder = proxy.configure_http_client(builder);
             builder.build()?
         };
@@ -138,3 +138,7 @@ fn prepare_request(
 
     request_builder.headers(content_headers).body(body)
 }
+
+#[cfg(test)]
+#[path = "../tests/unit/anonymous.rs"]
+mod tests;
