+++ libparsec/crates/client_connection/src/invited_cmds.rs
@@ -40,7 +40,7 @@ impl InvitedCmds {
         proxy: ProxyConfig,
     ) -> anyhow::Result<Self> {
         let client = {
-            let builder = reqwest::ClientBuilder::default();
+            let builder = reqwest::ClientBuilder::default().user_agent(crate::CLIENT_USER_AGENT);
             let builder = proxy.configure_http_client(builder);
             builder.build()?
         };
@@ -152,3 +152,7 @@ fn prepare_request(
 
     request_builder.headers(content_headers).body(body)
 }
+
+#[cfg(test)]
+#[path = "../tests/unit/invited.rs"]
+mod tests;
